# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API pendeteksi maqam musik Arab \u0026 Timur Tengah untuk komunitas banjari.",
        "title": "MaqamDetector API",
        "contact": {
            "name": "Anas (Cypress Consulting)"
        },
        "license": {
            "name": "ISC"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/maqamat/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "maqam"
                ],
                "summary": "Detail maqam",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Maqam ID (e.g. hijaz)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MaqamDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "dto.MaqamDetailResponse": {
            "type": "object",
            "properties": {
                "emotion_tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "example_songs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "interval_description": {
                    "type": "string"
                },
                "name_arabic": {
                    "type": "string"
                },
                "name_indonesia": {
                    "type": "string"
                },
                "name_latin": {
                    "type": "string"
                },
                "tips_aransemen": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}